+++ dwertheimer.TaskAutomations/changelog.md
@@ -8,6 +8,7 @@
 ## [2.11.0] (@dwertheimer)
 
 ### Added
+- Task Sorting: Skip Done/Cancelled; Include Checklist in Sorting
 - React Overdue task processing view v1
 - Overdue: change to checklist type